fix shopping cart up down quantity
