--- Views/ShoppingCart/AddShoppingCart.cshtml
@@ -1,5 +1,0 @@
-﻿@*
-    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-*@
-@{
-}
