+++ Views/ShoppingCart/ListCart.cshtml
@@ -33,9 +33,9 @@
                                     </td>
                                     <td class="col-sm-1 col-md-1" style="text-align: center">
                                         <div style="display:inline-flex">
-                                            <a asp-action="UpdateCartDown" asp-route-id="@item.product.id" asp-controller="Cart">-</a>
+                                            <a asp-action="UpdateCartDown" asp-route-id="@item.product.id" asp-controller="ShoppingCart">-</a>
                                             <p>@item.quantity</p>
-                                            <a asp-action="UpdateCartUp" asp-route-id="@item.product.id" asp-controller="Cart">+</a>
+                                            <a asp-action="UpdateCartUp" asp-route-id="@item.product.id" asp-controller="ShoppingCart">+</a>
                                         </div>
 
                                     </td>
